feat: Update player 2 public key display in gameDetails.js
The code changes in `gameDetails.js` modify the display of player 2's public key in the game details page. The public key is now displayed as a hyperlink, linking to the user's history games page. If the public key is "----", it is displayed as is without a hyperlink. This change improves the user experience by providing a convenient way to access the user's history games.